Add checks for the existence of some optional parameters.
We raise exceptions in config.create_circ_matrix_matchpoint.sql and in
config.create_hold_matrix_matchpoint.sql if we lookup an optional
value in the database and that lookup returns NULL.  We are assuming
that the user typoed the intended value.

TODO: Add checks for some of the other fields that are passed directly
to the insert queries at the end.

Signed-off-by: Jason Stephenson <jstephenson@mvlc.org>